Move events into an event package
